HW13 is complete

+++ hw12_13_14_15_calendar/internal/storage/entity/event.go
@@ -1,12 +1,12 @@
 package entity
 
 type Event struct {
-	ID           int    `json:"id" db:"id"`
-	UserID       int    `json:"userId" db:"user_id"`
+	ID           int64  `json:"id" db:"id"`
+	UserID       int64  `json:"userId" db:"user_id"`
 	Title        string `json:"title" db:"title"`
 	EventTime    string `json:"eventTime" db:"event_time"`
-	DurationMin  int    `json:"durationMin" db:"duration_min"`
-	NoticeBefore int    `json:"noticeBefore" db:"notice_before"`
+	DurationMin  int64  `json:"durationMin" db:"duration_min"`
+	NoticeBefore int64  `json:"noticeBefore" db:"notice_before"`
 	Description  string `json:"description" db:"description"`
 	CreatedAt    string `json:"createdAt" db:"created_at"`
 }
